Added ability to pass in options for RestClient
The RestClient can take a lot of options, and the low-level CouchRest::Database
can make use of them as params to it's methods, but there's NO way to get them
into the Database in order for them to be used. What we did was to make it
possible to set these options at the database level in the top-level public
API, and then save them with that database.